Move Addon.compatible_version to legacy_api.utils.find_compatible_version
